+++ src/pages/hypoteka.tsx
@@ -1,9 +1,9 @@
 import React from 'react'
 import Link from 'next/link'
 import { Form } from 'formik'
-import { BooleanRadio, FormWrapper, Input } from '../components/FormComponents'
+import { BooleanRadio, FormWrapper, Input } from '../plugins/_components/form/FormComponents'
 import { FormErrors, MortgageUserInput } from '../types/PageUserInputs'
-import { ErrorSummary } from '../components/ErrorSummary'
+import { ErrorSummary } from '../plugins/_components/form/ErrorSummary'
 import { numberInputRegexp } from '../lib/utils'
 import { Page } from '../components/Page'
 import { mortgageInitialValues } from '../lib/initialValues'
