+++ src/pages/odklad/pokracovat.tsx
@@ -3,7 +3,7 @@ import getConfig from 'next/config'
 import { Page } from '../../components/Page'
 import { PostponeUserInput } from '../../types/PostponeUserInput'
 import { convertPostponeToXML } from '../../lib/postpone/postponeConverter'
-import { RedirectField, RedirectForm } from '../../components/RedirectForm'
+import { RedirectField, RedirectForm } from '../../plugins/_components/form/RedirectForm'
 import { setDate, toBase64 } from '../../lib/utils'
 import { TAX_YEAR } from '../../lib/calculation'
