+++ src/pages/pokracovat.tsx
@@ -3,7 +3,7 @@ import getConfig from 'next/config'
 import { Page } from '../components/Page'
 import { TaxForm } from '../types/TaxForm'
 import { convertToXML } from '../lib/xml/xmlConverter'
-import { RedirectField, RedirectForm } from '../components/RedirectForm'
+import { RedirectField, RedirectForm } from '../plugins/_components/form/RedirectForm'
 import { setDate, toBase64, formatCurrency } from '../lib/utils'
 import { buildSummary } from '../lib/calculation'
 import { Summary } from '../types/Summary'
