+++ src/pages/vysledky.tsx
@@ -3,7 +3,7 @@ import { formatCurrency } from '../lib/utils'
 import { TaxFormUserInput } from '../types/TaxFormUserInput'
 import { Page } from '../components/Page'
 import Decimal from 'decimal.js'
-import { BackLink } from '../components/BackLink'
+import { BackLink } from '../plugins/_components/links/BackLink'
 import Link from 'next/link'
 import { buildSummary } from '../lib/calculation'
 import { TAX_YEAR } from '../lib/calculation'
