+++ src/pages/zamestnanie.tsx
@@ -1,12 +1,12 @@
 import React from 'react'
 import { Form } from 'formik'
-import { BooleanRadio, FormWrapper, Input } from '../components/FormComponents'
+import { BooleanRadio, FormWrapper, Input } from '../plugins/_components/form/FormComponents'
 import { EmployedUserInput, FormErrors } from '../types/PageUserInputs'
 import { numberInputRegexp } from '../lib/utils'
-import { ErrorSummary } from '../components/ErrorSummary'
+import { ErrorSummary } from '../plugins/_components/form/ErrorSummary'
 import { Page } from '../components/Page'
 import { employmentUserInputInitialValues } from '../lib/initialValues'
-import { BackLink } from '../components/BackLink'
+import { BackLink } from '../plugins/_components/links/BackLink'
 import { TAX_YEAR } from '../lib/calculation'
 
 const Zamestnanie: Page<EmployedUserInput> = ({
