+++ src/plugins/_components/form/Details.tsx
