+++ src/plugins/main/_layout/PrivacyPolicyLink.tsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import { ExternalLink } from './ExternalLink'
+import { ExternalLink } from '../../_components/links/ExternalLink'
 
 export const PrivacyPolicyLink = () => (
   <ExternalLink href="https://navody.digital/ochrana-osobnych-udajov">
